Add stack depth check to feature iterator
This adds a safety guard for malformed expressions.
For example, self-referencing virtual fields. It's impossible to detect all possible triggers for this during preparation.

A simple hand-crafted example to illustrate an impossible to detect situation:

A layer with one field, let's call it "v", and on feature with the value "virtual" as value for "v".

A virtual field with the name "virtual" and the definition `eval("v")` will cause an infinite recursion.

This patch will simply check the recursion depth and bail out after 255 nested iterators.